requirements.txt
Pillow==10.4.0
requests==2.32.3
Flask==3.0.3
ttkthemes==3.2.2
